register humanized association names if different from link_fields
e.g. if a custom relationship is created in the Studio between Tasks and Documents, the corresponding link_field will be
Since a human would refer to the association as  (e.g. by calling task.documents), the humanized name of the association is registered in the Class

This is done in addition to registering a method with the link_field name, to maintain backward compatibility